+++ src/components/Node.tsx
@@ -1,5 +1,5 @@
 import React, { useContext, useMemo } from "react";
-import { useProperty, useNode } from "../hooks";
+import { useProperty, useNode, useOffsetPosition, useIntegerPosition } from "../hooks";
 import { Prop, Position } from "../types";
 import { NodeContext } from "../context";
 
@@ -13,14 +13,23 @@ import { NodeContext } from "../context";
 type NodeProps = {
   children: React.ReactNode;
   pos?: Prop<Position>;
+  offset?: Prop<Position>;
+  rounded?: boolean;
   timeScale?: Prop<number>;
 }
 
-export const Node: React.FC<NodeProps> = ({ children, timeScale, ...props }) => {
+export const Node: React.FC<NodeProps> = ({ children, timeScale, rounded, ...props }) => {
   const parent = useContext(NodeContext);
   const pos = useProperty<Position>(props.pos || parent.pos || [0, 0]);
+  const offsetPos = useOffsetPosition(pos, props.offset || [0, 0]);
+  const roundedPos = useIntegerPosition(offsetPos);
   const node = useNode({ timeScale });
-  const context = useMemo(() => ({ ...node, debug: parent.debug, pos }), [node, parent.debug, pos]);
+
+  const context = useMemo(() => ({
+    ...node,
+    debug: parent.debug,
+    pos: rounded ? roundedPos : offsetPos,
+  }), [node, parent.debug, pos]);
 
   return <NodeContext.Provider value={context}>{children}</NodeContext.Provider>;
 };